+++ src/bindings/index.js
@@ -1,0 +1,11 @@
+/**
+ * Created by Paul on 9/7/2016.
+ */
+
+import opt from './opt';
+import get from './get';
+import key from './key';
+import path from './path';
+import val from './val';
+
+export {opt, get, key, path, val};
